Move `console.log` rerouting into own module
This makes the daemon.js file a little cleaner. As suggested by
@lencioni in code review.
