+++ lib/daemon.js
@@ -1,8 +1,6 @@
-import fs from 'fs';
 import os from 'os';
 import path from 'path';
 import readline from 'readline';
-import util from 'util';
 
 import commander from 'commander';
 import loglevel from 'loglevel';
@@ -11,23 +9,10 @@ import loglevelMessagePrefix from 'loglevel-message-prefix';
 import Configuration from './Configuration';
 import Importer from './Importer';
 import WatchmanFileCache from './WatchmanFileCache';
+import rerouteConsoleLog from './rerouteConsoleLog';
 
 const pathToLogFile = path.join(os.tmpdir(), 'importjs.log');
-const logStream = fs.createWriteStream(pathToLogFile, {
-  flags: 'w',
-});
-
-/* eslint-disable no-console */
-const originalConsoleLog = console.log;
-console.log = (...args) => {
-  logStream.write(`${util.format.apply(null, args)}\n`);
-};
-console.trace = console.log;
-console.debug = console.log;
-console.info = console.log;
-console.warn = console.log;
-console.error = console.log;
-/* eslint-enable no-console */
+const originalConsoleLog = rerouteConsoleLog(pathToLogFile);
 
 const commandsToFunctionNames = {
   fix: 'fixImports',
